Refactor the code into separate files.
Each command is now grouped into separate files to make it more maintainable.
The grouping is based on the documentation in redis.io website.
